+++ Gruntfile.js
@@ -1,4 +1,3 @@
-<?php
 // This file is part of Moodle - http://moodle.org/
 //
 // Moodle is free software: you can redistribute it and/or modify
@@ -14,31 +13,35 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
+/* eslint-env node */
+
 /**
- * access capabilities describe
+ * Local Grunt tooling for report_overviewstats.
+ *
+ * Lints this plugin's CSS with the same rules Moodle core's own Gruntfile
+ * enforces (see .stylelintrc.json), so issues that moodle-plugin-ci's
+ * "grunt" CI step would flag can be caught locally without a full Moodle
+ * checkout.
  *
  * @package report_overviewstats
- * @category access
  * @author DualCube <admin@dualcube.com>
- * @copyright 2013 David Mudrak <david@moodle.com>
  * @copyright 2023 DualCube <admin@dualcube.com>
  * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+module.exports = function(grunt) {
+    grunt.loadNpmTasks('grunt-stylelint');
 
-defined('MOODLE_INTERNAL') || die();
+    grunt.initConfig({
+        stylelint: {
+            css: {
+                options: {
+                    configFile: '.stylelintrc.json',
+                    quietDeprecationWarnings: true,
+                },
+                src: ['styles.css'],
+            },
+        },
+    });
 
-$capabilities = [
-
-    'report/overviewstats:view' => [
-        'riskbitmask' => RISK_PERSONAL,
-        'captype' => 'read',
-        'contextlevel' => CONTEXT_COURSE,
-        'archetypes' => [
-            'teacher' => CAP_ALLOW,
-            'editingteacher' => CAP_ALLOW,
-            'manager' => CAP_ALLOW,
-        ],
-
-        'clonepermissionsfrom' => 'coursereport/stats:view',
-    ],
-];
+    grunt.registerTask('default', ['stylelint']);
+};