+++ lang/en/report_overviewstats.php
@@ -14,20 +14,23 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
- /**
-  * Provides a list of strings for the plugin
-  *
-  * @package report_overviewstats
-  * @category string
-  * @author DualCube <admin@dualcube.com>
-  * @copyright 2023 DualCube <admin@dualcube.com>
-  * @copyright based on work by 2013 David Mudrak <david@moodle.com>
-  * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
-  */
+/**
+ * Provides a list of strings for the plugin
+ *
+ * @package report_overviewstats
+ * @category string
+ * @author DualCube <admin@dualcube.com>
+ * @copyright 2013 David Mudrak <david@moodle.com>
+ * @copyright 2023 DualCube <admin@dualcube.com>
+ * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
 
 defined('MOODLE_INTERNAL') || die();
 
 $string['pluginname'] = 'Overview statistics';
+$string['chart-access'] = 'Course access';
+$string['chart-access-group'] = 'Course access ({$a})';
+$string['chart-access-perday'] = 'Per day';
 $string['chart-countries'] = 'User countries';
 $string['chart-countries-info'] = 'Number of different countries: {$a}';
 $string['chart-courses'] = 'Courses';
@@ -38,6 +41,7 @@ $string['chart-courses-percategory-coursesown'] = 'Courses (own only)';
 $string['chart-courses-sizes'] = 'Number of courses per size';
 $string['chart-courses-sizes_help'] = 'Displays the distribution graph of number of activities per course. That is, how many courses are there with 0-4 activities, 5-9 activities, 10-14 activities etc.';
 $string['chart-enrolments'] = 'Enrolled users';
+$string['chart-enrolments-group'] = 'Enrolled users ({$a})';
 $string['chart-enrolments-month'] = 'Last month';
 $string['chart-enrolments-year'] = 'Last year';
 $string['chart-langs'] = 'User preferred languages';
@@ -45,7 +49,10 @@ $string['chart-langs-info'] = 'Number of different languages: {$a}';
 $string['chart-logins'] = 'Users logging in';
 $string['chart-logins-perday'] = 'Per day';
 $string['overviewstats:view'] = 'View overview statistics';
-$string['privacy:metadata'] = 'The overview statistics plugin does not store any personal data.';
+$string['privacy:metadata:core_enrol'] = 'The overview statistics plugin reads the number of currently enrolled users, via the core enrolment subsystem, to build the enrolments chart. It does not store this data itself.';
+$string['privacy:metadata:core_group'] = 'The overview statistics plugin reads group membership, via the core group subsystem, to filter the enrolments chart by group. It does not store this data itself.';
+$string['privacy:metadata:core_user'] = 'The overview statistics plugin reads user country and language fields, via the core user subsystem, to build the countries and languages charts. It does not store this data itself.';
+$string['privacy:metadata:logstore'] = 'The overview statistics plugin reads login, course access, and enrolment log events, via the installed log store, to build historical charts. It does not store this data itself.';
 $string['user-numbers'] = 'Number of users';
 $string['course-numbers'] = 'Number of course';
 $string['enrolled'] = 'Enrolled';