+++ lib.php
@@ -19,8 +19,8 @@
  *
  * @package report_overviewstats
  * @author DualCube <admin@dualcube.com>
+ * @copyright 2013 David Mudrak <david@moodle.com>
  * @copyright 2023 DualCube <admin@dualcube.com>
- * @copyright based on work by 2013 David Mudrak <david@moodle.com>
  * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
@@ -32,14 +32,20 @@
  * @param stdClass $context The context of the course
  */
 function report_overviewstats_extend_navigation_course($navigation, $course, $context) {
-    // There is no data for the frontpage course
+    // There is no data for the frontpage course.
     if ($course->id == SITEID) {
         return;
     }
 
     if (has_capability('report/overviewstats:view', $context)) {
         $url = new moodle_url('/report/overviewstats/index.php', ['course' => $course->id]);
-        $navigation->add(get_string('pluginname', 'report_overviewstats'), $url, navigation_node::TYPE_SETTING,
-            null, null, new pix_icon('icon', '', 'report_overviewstats'));
+        $navigation->add(
+            get_string('pluginname', 'report_overviewstats'),
+            $url,
+            navigation_node::TYPE_SETTING,
+            null,
+            null,
+            new pix_icon('icon', '', 'report_overviewstats')
+        );
     }
 }